+++ web/global.d.ts
@@ -1,4 +1,5 @@
 import {
+    GitCommandStatus as GitCommandStatusX,
     GitCommitNode as GitCommitNodeX,
     GitGraphViewSettings as GitGraphViewSettingsX,
     RequestMessage as RequestMessageX,
@@ -6,6 +7,7 @@ import {
 } from "../out/types";
 
 declare global {
+    type GitCommandStatus = GitCommandStatusX;
     type GitCommitNode = GitCommitNodeX;
     type GitGraphViewSettings = GitGraphViewSettingsX;
     type RequestMessage = RequestMessageX;